........
  r696501 | josh | 2008-09-17 15:35:09 -0700 (Wed, 17 Sep 2008) | 1 line

  Small tweak to NamePtg and Ref3DPtg
........


git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@696831 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/HSSFName.java
@@ -48,7 +48,7 @@ public class HSSFName implements Name {
      * @return sheet name, which this named range referred to
      */
     public String getSheetName() {
-        short indexToExternSheet = _definedNameRec.getExternSheetNumber();
+        int indexToExternSheet = _definedNameRec.getExternSheetNumber();
 
         return _book.getWorkbook().findSheetNameFromExternSheet(indexToExternSheet);
     }
